profiler_worker: move all localnet related logic from profiler to helpers/localnet_handler.py
